Seperate AstraSensors.h into .h and .cpp
